docs/planning/greenlang-2030-vision/agent_foundation/messaging/requirements.txt
# GreenLang Messaging System Dependencies
# Production-ready message broker for distributed agents

# Core Dependencies
redis[hiredis]>=5.0.0  # Redis client with C parser for performance
pydantic>=2.5.0  # Data validation and serialization
pyyaml>=6.0  # YAML configuration parsing

# Optional: High-performance serialization
msgpack>=1.0.0  # Binary serialization (optional, faster than JSON)

# Optional: Kafka support (for scale deployment)
# Uncomment if using Kafka broker
# kafka-python>=2.0.0
# confluent-kafka>=2.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Monitoring (optional)
prometheus-client>=0.18.0  # Prometheus metrics

# Development
black>=23.0.0  # Code formatting
ruff>=0.1.0  # Linting
mypy>=1.7.0  # Type checking
